Refactor reconciliation operations (CS1).
Refactor the code for reconciliation in order to use our latest code standards and guidelines.
Added the support to download files!, now we can use the download method to retrieve CSV files.
Removed wiremock test dependency as currently we don't plan to use it. Evaluating in a future a case to use.
Small refactor on api path builder to avoid the repeated code on each class.